requirements.txt
flask==2.0.1
flask-wtf==0.15.1
python-binance==1.0.16
pandas==1.3.5
numpy
plotly==5.10.0
dash==2.6.2
dash-bootstrap-components==1.2.1
python-dotenv==0.21.0
